(#9, #17): Refactor main into learning context.
